+++ khipu-eth/src/main/scala/khipu/domain/Account.scala
@@ -2,12 +2,12 @@ package khipu.domain
 
 import akka.util.ByteString
 import khipu.Hash
+import khipu.UInt256
 import khipu.crypto.kec256
 import khipu.network.p2p.messages.PV63.AccountImplicits
 import khipu.rlp
 import khipu.rlp.RLPImplicits._
 import khipu.trie.ByteArraySerializable
-import khipu.vm.UInt256
 
 object Account {
   // 56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421
